backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
langchain-google-genai
sqlalchemy
psycopg2-binary
python-jose
websockets
python-dotenv
pydantic-settings